remove redundant folder structure;
