+++ app/app.go
@@ -33,7 +33,7 @@ import (
 	"github.com/ChainSafe/sygma-relayer/chains/evm/executor"
 	"github.com/ChainSafe/sygma-relayer/chains/evm/listener"
 	"github.com/ChainSafe/sygma-relayer/chains/substrate"
-	substrate_bridge "github.com/ChainSafe/sygma-relayer/chains/substrate/calls/pallets/bridge"
+	substrate_pallet "github.com/ChainSafe/sygma-relayer/chains/substrate/calls"
 	"github.com/ChainSafe/sygma-relayer/chains/substrate/client"
 	"github.com/ChainSafe/sygma-relayer/chains/substrate/connection"
 	substrateExecutor "github.com/ChainSafe/sygma-relayer/chains/substrate/executor"
@@ -209,7 +209,7 @@ func Run() error {
 					panic(err)
 				}
 				client := client.NewSubstrateClient(conn, &keyPair, config.ChainID)
-				bridgePallet := substrate_bridge.NewBridgePallet(client)
+				bridgePallet := substrate_pallet.NewPallet(client)
 
 				depositHandler := substrate_listener.NewSubstrateDepositHandler()
 				depositHandler.RegisterDepositHandler(message.FungibleTransfer, substrate_listener.FungibleTransferHandler)
